+++ src/coverage-reporter/coverage-calculation-helpers.ts
