+++ tests/coverage-logger.test.ts
